refactor : 그룹 리팩토링
